Start reworking client
